minimal name change for consistency
I replaced cog_dict with gene_clusters_dict and checkm_dict with
genome_completion_dict since I believe these are more accurate names
that will not lead to a lot of confusion later. for instance, anvi'o
does not compute genome completeness using checkm, and reflecting it in
the codebase is I think a better practice.

The rest of the changes in anvi-run-motupan works with these variable
names.