a new heuristic for missing completion estimates
The idea here is to use existing completion estimates, and recover those
that are missing based on what we know about, rather than genome lenghts
alone. The new code identifies the most highly complete genomes in the
set, takes the average of the lengths of those genomes, and then
assignts completion estimates to those that are missing any estimates. I
tested it with a few pangenomes, and I got pretty good results.